Add new access and conversion functionality for Value
- Conversion to / from `std::vector<T>` iff possible
- Construction from unsigned
- Unmarshal of `std::istream` input
- Enhanced access function allowing a chain of keys (`std::vector<std::string>`) as parameter
- Enhanced access function including a conversion to specified type iff possible

+++ test/test_marshal.cpp
@@ -8,7 +8,7 @@ static std::string _readStream(std::ifstream *pInfile) {
   assert(pInfile->is_open());
 
   std::string ret;
-  ret.resize(pInfile->tellg());
+  ret.resize(static_cast<size_t>(pInfile->tellg()));
   pInfile->seekg(0, std::ios::beg);
   pInfile->read(&ret[0], ret.size());
   pInfile->close();
